Functional Tests: Adapt psbt test to 80'000'000 supply limit
Change psbt functional test cases where the max money range checks are
made. Reproducing these cases was much more annoying than anticipated:

For the input increase, recompile with a higher max money range and
subsidy to increase the amounts in the functional test, then grap one of
the encoded psbts from the functional tests either with a debugger or
print statement.

For the output increase, change one of the psbt test to use a higher
amount, then grab a finalized psbt with that amount.